py/services/orders/pyproject.toml
[tool.poetry]
name = "orders"
version = "0.1.0"
description = ""
authors = ["Aleksey Dalekin <adalekin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.9"
connexion = "2.9.0"
connexion-buzz = "0.2.1"
environs = "9.3.2"
flask-idempotency = {path = "../../lib/flask-idempotency"}
flask-marshmallow = "0.14.0"
Flask-Migrate = "3.1.0"
Flask-SQLAlchemy = "2.5.1"
gunicorn = "20.0.4"
marshmallow = "3.13.0"
marshmallow-enum = "1.5.1"
marshmallow-sqlalchemy = "0.26.1"
psycopg2 = "2.9.1"
sentry-sdk = {version = "0.7.10", extras = ["flask"]}
SQLAlchemy = "1.4.22"
talepy = "0.8.0"

[tool.poetry.dev-dependencies]
coverage = "5.5"
pytest = "6.2.4"
pytest-env = "0.6.2"
pytest-mock = "3.6.1"
requests-mock = "1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
